페이징 프론트 수정

+++ src/main/java/com/multicampus/topicsation/dto/pageDTO/PageResponseDTO.java
@@ -27,7 +27,7 @@ public class PageResponseDTO {
         this.dataPerPage = pageReqeustDTO.getDataPerPage();
         this.pagePerOnce = pageReqeustDTO.getPagePerOnce();
 
-        this.totalPageCount = (int)Math.ceil(totalDataCount / dataPerPage);
+        this.totalPageCount = (int)Math.ceil(totalDataCount /(double)dataPerPage);
         this.startPage = (int)Math.ceil(currentPage /(double)pagePerOnce) * pagePerOnce - pagePerOnce + 1;
         this.endPage = startPage + pagePerOnce - 1;
         this.endPage = endPage > totalPageCount ? totalPageCount: endPage;
